# Conflicts:
#	DelvUI/Interface/GeneralElements/CastbarConfig.cs
#	DelvUI/Interface/GeneralElements/LabelConfig.cs
#	DelvUI/Interface/GeneralElements/LabelHud.cs
#	DelvUI/Interface/GeneralElements/PrimaryResourceConfig.cs
#	DelvUI/Interface/GeneralElements/UnitFrameConfig.cs
#	DelvUI/Interface/StatusEffects/StatusEffectsListConfig.cs
#	DelvUI/Plugin.cs

+++ DelvUI/Config/Attributes/ConfigTypeAttributes.cs
@@ -296,7 +296,7 @@ namespace DelvUI.Config.Attributes
             {
                 field.SetValue(config, intVal);
 
-                TriggerChangeEvent<float>(config, field.Name, intVal);
+                TriggerChangeEvent<int>(config, field.Name, intVal);
 
                 return true;
             }
